PsaDriver: Update the usage of internal structure for keys
Direct private structure should not be used for key attributes.
Fix it to use PSA provided functions.

Domain paramaer is no longer supported. Remove it's check.

There is one place where we actually use the value of domain paaramter
if available. That is commented under a MACRo and can be enabled once
parameter passing works with PSA driver wrapper layer in later versions.

Signed-off-by: Ruchika Gupta <ruchika.gupta_1@nxp.com>